requirements.txt
wheel
nbresult
colorama
ipdb
ipykernel
holidays
yapf
matplotlib
pygeohash
pytest
seaborn
numpy==2.2.6
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2
google-cloud-bigquery
google-cloud-storage==2.14.0
google-api-core==2.8.2
googleapis-common-protos==1.56.4
protobuf==3.19.6
h5py==3.10.0
db-dtypes
pyarrow
statsmodels==0.14.0
torch
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

# prevent bq 3+ db-dtypes package error when running training from bq

mlflow==2.1.1
prefect==2.19.2

python-dotenv
psycopg2-binary

# API
fastapi
uvicorn
requests
# tests
httpx<0.28
pytest-asyncio
ipykernel

requests-cache
retry-requests
pandas-gbq
